requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
email-validator
sqlalchemy
alembic
python-multipart
python-jose[cryptography]
python-dotenv
httpx
pytest-asyncio
black
isort
flake8
mypy
passlib[bcrypt]
psycopg2-binary